Fix #248 Lifeline Header
Add an API to the LayoutHelper to adjust proposed
bounds for an existing shape in the diagram.  Use this
new API to constrain the moving of lifeline headers
to the horizontal dimension (including visual feed-back).
Editing the size of a lifeline header likewise is constrained
to the horizontal dimension (otherwise the lifeline body
would be moved vertically).

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>