Issue #295 Show notation in Logical Model View
Show the diagram notation views also in the Logical Model View
for ease of debugging, to inspect the properties of the notation.
Update the selection handling to select the notation view of the
selected edit-part instead of the logical model element (which
then is an ancestor it the view in the tree).

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>