Move layout-related magic numbers to layout constraints
The LayoutConstraints is heavily overloaded to account for allowing to
conveniently obtain the constants with a view, for a part of the view,
or only a view type. Most of the edit parts and policies should query
the constants with their view, to enable in the future taking e.g.
styles into account.

This change for now only centralizes the magic numbers of edit parts and
edit policies. There are several more magic numbers in the figures,
which have to be migrated. For those, I suggest to provide them from the
edit parts instantiating the view.