Lifeline resize: Ensure strict compound
Introduce a NoopCommand as alternative to the IdentityCommand
API that lets logical model commands compose themselves using
the pessimistic strict CompoundModelCommand.  This ensures that
in commands such as the ResizeHorizontallyCommand, chains
beginning on the no-op will not lose the benefit of the strict compound
that lets each command in sequence calculate its parameters on
the basis of the results of the previous command.

Also fix commands to resize/move the rightmost lifeline not being
executable because the move visitor doesn’t compute anything
as there are no lifelines further to the right for it to operate on.