Issue #73 Execution start/finish on message ends
Update alignment constraints and feed-back of message
creation to enforce horizontality of non-asynchronous
messages by recalculating the sending end instead of the
receiving end because the receiving end is generally
more significant.

Implement a magnet mechanism to snap connections to
the ends of an execution specification.  Upon creation of
a message connecting to either end of an execution
specification, replace the occurrence at that end by the
message end.

Register a palette provider that injects magnet manager
support into the selection tool and connection creation
tools to support user override of the magnet snapping
behaviour when drawing and moving connections, using
the same keyboard modifier as for suppression of the
snapping of shapes.

Update a create message test to account for messages
anchored on the beginning of an execution specification
correctly remaining attached when the execution moves.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>

# Conflicts:
#	plugins/org.eclipse.papyrus.uml.interaction.model/src/org/eclipse/papyrus/uml/interaction/internal/model/commands/InsertMessageCommand.java
#	plugins/org.eclipse.papyrus.uml.interaction.model/src/org/eclipse/papyrus/uml/interaction/internal/model/spi/impl/DefaultDiagramHelper.java
#	plugins/org.eclipse.papyrus.uml.interaction.model/src/org/eclipse/papyrus/uml/interaction/model/spi/DiagramHelper.java