Fix issue #343 Restrict message end movement
Implement the requirement to not allow movement of a message end
to result in the other end being moved also to maintain semantic
integrity.  Thus, only the end grabbed by the user can be moved and
then only if the resulting message (the other end always staying put)
would be valid.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>

# Conflicts:
#	plugins/org.eclipse.papyrus.uml.diagram.sequence.runtime/src/org/eclipse/papyrus/uml/diagram/sequence/runtime/internal/edit/policies/AbstractSequenceGraphicalNodeEditPolicy.java