Issue #343: Insertion problems
Fix the semantic insertion order of interaction fragments
in the case of creating an execution specification.

Includes fixing a test that didn’t account for tear-down of
the logical model when the model is changed.  Also adding
the Papyrus di resources to git for ease of reference to the
notation in a Papyrus run-time workbench.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>