Support auto-creating executions when adding sync messages
This change addresses the following requirements of issue #40:
* An execution specification SHALL be created in the lifeline with a
predefined height.
* The start occurrence specification SHALL be set to the message
occurrence specification for the receive event.
* Optionally (determined by user configuration), a Reply message will be
created between the same ends, but in opposite direction and if so, the
finish occurrence specification of the execution specification SHALL
correspond to the send event message occurrence specification of the
reply message.

Actually, in this change we also make the entire creation of the
execution optional based on a preference. This is also leveraged to
avoid breaking existing tests that don't assume that an execution is
created.

Still open after this change is the following item:
* The user will be asked for the type of the execution specification,
BehaviorExecutionSpecification or ActionExecutionSpecification.

Signed-off-by: Philip Langer <planger@eclipsesource.com>