Issue #76: Adjust Y calculations in logical model
The Logical Model depends on a plugged-in LayoutHelper
to determine X and Y positioning of notation views in the
diagram without reference to the GEF presentation layer.
However, this calculation does not account for the platform-
specific font metrics of the interaction name label, and
that error blocks testing of a fix for Issue #76.

So, add a pluggable FontHelper component that the
LayoutHelper can use to compute the height of the
interaction label according to the font that it uses to
render text.  This must be injected by a UI-aware
bundle because the Logical Model is a headless API.

Other consequential updates to tests:
- can un-ignore a test that was ignored because of this Y error
- some layout test assertions have to be adjusted, especially
tests on a model that doesn’t have interaction name labels at
all (this model pre-dating that detail of the implementation)
- other tests that create elements too close to other elements
now are affected by the padding feature, so they need to
account for that

In regression testing it was observed that the message
reconnection constraints tests as written didn’t do what
they were intended to do and so always skipped on an
assumption violation; these are now fixed.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>