Prefs rule doesn’t support multiple usage
When multiple test suites use the prefs rule, because it sets
the preference on construction, all preference setting is done
when the test suites are created.  Consequently, after the
first suite runs its afters, all preferences will have reverted
to the default and all subsequent tests using the rule will
not see it have any effect.  This commit postpones the
setting of preferences to the before execution and also
ensures correct reversion of preferences in case of nested
or multiple application of the rule.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>