Correct y-position reference for computing direct pre- and successor
We were basing the computation of the latest and following element on
the y position that has already been adjusted to keep minimum padding.
This could lead to scenarios, where we inadvertantly start nudging one
element too far, when the insertion is very close to the next element
(close enough so that the added padding moved it above the Y of the next
element).