Issue #52 Primitive setCovered operation
Implement logical model operations
 - MOccurrence::setCovered(MLifeline, OptionalInt)
 - MExecution::setOwner(MLifeline, OptionalInt)

operation to create a command that changes the semantic
lifeline coverage of an occurrence or execution, with an
option to set a new Y position on the lifeline.  These APIs
are now used in the edit-policy to effect the reconnection
of messages, either to a new lifeline at the same Y position
or to another Y position on the current lifeline

When the re-targeted message end is a destruction occurrence,
ensure that its view is moved to the new target lifeline. The
message end remains attached to the destruction X shape.

If a message end being moved to another lifeline is attached
to an execution specification, move that execution, also.
Further, when moving an execution specification, move its
start and finish occurrences, whatever they may be.

Includes a reworking of the graphical node edit-policy handling
message reconnection, to delegate the commands effecting
both the visuals and the semantics of reconnection to the
logical model.  This is necessary to ensure that we don’t
attempt redundant/cyclical updates of the dependencies
of, e.g., execution specifications.  Besides that this more
closely aligns to the goal of having all modifications
implemented in the Logical Model, with the diagram
delegating to it.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>