Issue #52 Handle occurrences in executions
Implement support for pulling along other occurrences spanned
by an execution that is being moved to another lifeline (which in
turn is because ither its start or its finish is being moved).

This includes new APIs in the *Logical Model* to query the
occurrences covering an `MLifeline` and spanned by an
`MExecution`.

Handling the case where an end of a message is brought
to the same lifeline as the other end, thus making it a
self-message, is handled in a fairly primitive way and needs
further elaboration.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>