Issue #52 Fix JUnit test failure on build server
A JUnit test fails on the build server because the automation
method of creating a component of the test fixture doesn’t
work the same in the Linux environment as on Mac.  Since this
test fixture isn’t what the test is trying to test, force its creation
by direct manipulation of commands in the diagram.

This exposed a latent issue in the Logical Model implementation
of the GMF request/command protocol in which view descriptors
for CreateViewAndElementCommands were not being completed
by injection of the created view.  That is fixed in order to
implement the new test framework.

And, now that the view descriptor from the request gets its view,
the layout edit-policy for the interaction has to be aware that of
the constraint on positioning and sizing lifeline heads (previously
it didn’t matter because the view in the descriptor was null and
so the SetBoundsCommand was just ineffectual).

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>