Issue #52 Fix reparent-and-place errors
When positioning an element, the relative location calculation
must be deferred in case the element is also be re-parented to
a parent that is at a different location (especially in the Y dimension)
than the former parent.

Also, ensure that we don’t try to pad the create event of a
lifeline or whatever other event is the first in case that the new
parent lifeline has no occurrences below the point of insertion.

Finally, ensure that dependencies of the initial element to be
moved to another lifeline can override what was previously
determined on that lifeline to require nudging, to record that
nothing requires nudging because (for example) all existing
elements on that lifeline are now subsumed into an execution
specification that was moved to it.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>