Issue #73 Deterioration of connection feedback
The GMF adapters for CSS force refresh of the diagram on
certain changes to semantic model elements.  One such
case is removal of a contained object, which occurs routinely
in the preparation of pessimistic compound model commands.
That disrupts message connection feed-back by refreshing
the message edit-part, which then resets the connection
away from the mouse pointer where the user is directing it.

Install a custom CSS adapter that knows when not to refresh
the diagram.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>