Fix #65 Ensure correct fragment ordering
Ensure that when start/finish occurrences of executions, or
message ends, are moved visually relative to other fragments
(with the keyboard modifier), the semantic ordering if fragnents
is updated accordingly.

Incidentally noticed, in adding externalized text for commands,
that an existing command didn’t use the EMF pattern already
implemented, so refactored that.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>