Fix #70 JUnit test failures
Once again, the Ctrl modifier key for semantic re-ordering
override interfered with standard GEF editing behaviour,
which on Linux and Windows clones the object being
dragged (on Mac it’s the Option modifier).

So, in the drag tracker, disable the clone behaviour that
we don’t really want in this diagram anyways.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>