Auto-correct lifeline of an executions' reply message
When re-orienting a message that has an end being the start or finish of
an execution, this change also re-orients the opposite accordingly, so
that both eventually start/finish on the same lifeline.

With that, we ensure the validity of the following constraint:
If a reply message finishes the specification, this is, the finish
occurrence specification of the execution specification corresponds with
the send event message occurrence specification of the reply message,
the receive event of the message must be located in the same lifeline
than the send event of the message starting the execution specification,
if any.

Fix #42

Signed-off-by: Philip Langer <planger@eclipsesource.com>