Respect lifeline's offset for deciding whether to forbid re-orientation
When deciding whether to forbid a send/reply re-orientation of an
execution or not, we determine the element at or closest before the
y-position where the opposite message is going to be, in order to test
whether there is a destruction occurrence before.

However, the y-position that we are using turns out to be an absolute
y-position, so we need to take the lifeline's offset into account.

Signed-off-by: Philip Langer <planger@eclipsesource.com>