Fix #385 Too much padding on create message
Re-use the command employed in re-targeting create messages
for the initial create message scenario, too.  Also reuse the
common deferred padding mechanism for the padding from the
new lifeline head location.

In order to make the latter work, it is necessary also to fix a
latent problem in which the duplicate calculation of commands
for the “unspecified type” create requests is trips up the
dependency context’s tracking of unique command instances.
Also various problems in the handling of destruction occurrences
in nudging.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>