Issue #26 Excess nudging
Don’t nudge to maintain all of the whitespace before the elements following an
inserted message.  Rather, nudge by the amount of vertical space consumed
by the new message.

This is accomplished by unifying the nudging of elements for the new message
with the DeferredPaddingCommand pattern.  This, in turn, requires the logical
model to create its representation of the newly created message, including any
executions and replies also created.  Furthermore, correct padding also
requires recognizing and ignoring elements that have been deleted, so the
logical model is also updated to remove elements that are deleted and to
provide an exists() query.  Finally, this removal of logical model elements
requires alignment of usage and definition of the RemovalCommand protocol
with the CreationCommand protocol.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>