Issue #26 Request/reply constraint
Restore code removed that was necessary for the implementation
of the synchronous message request/reply constraint.  Add JUnit
tests covering that constraint to prevent another regression.  Also
fix the assumption of logical model identity in the implementation
of that constraint (only the identity of underlying UML model
elements may be relied upon).

Incidentally fix the notation of the model used for the JUnit test
to properly attach the request and reply messages to the execution
specification.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>