Fix #26 code review issues
Fix issues identified in code review, covered by new tests for each of:
- nudging can push elements up above the top of lifeline
- moving a message doesn’t nudge elements up
- stretching an execution doesn’t nudge elements down

One of the problems in the incremental construction of complex
layout commands is that normal cases (esp. create messages)
can produce temporary invalid layouts that are later corrected.
One such case is negative connection anchor distances, placing
messages above lifelines, which are then later nudged down
again.  To ensure that cases of such invalid layout result in
unexecutable commands, but only when not a temporary
condition, this change introduces a concept of “kill switch”.
This is a command that is appended to the composite after
all others at the top (edit-policy) level, which when evaluated
will be unexecutable unless it was previously countermanded
(such as by a command that corrects an invalid layout).

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>