Fix #439 Cannot delete a message
Fix the failure of a nudge operation that isn’t needed for
messages that are implicitly nudged by their attached
source and target views also being nudged, which is
needlessly unexecutable.

Also fix the SortSemanticsCommand that sporadically
fails on shuffles of the fragments that break the naïve
MoveCommand implementation.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>