Issue #389: Fix test assumption violations on Mac
Fix several new test assumption violations on Mac platform, along
with some compiler warnings in files changed for that reason.  Note
that one of the LifelineSwitchingUITest cases that was being skipped
on an assumption violation is now (accurately) showing a regression
on all platforms (a message that becomes a self-message now
doesn’t have any vertical extent where it does on the master branch).

Also fix the definition of some sub-suites of the LifelineSwitchingUITest
that were making wrong assertions/assumptions.  Those tests now
pass instead of skip.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>